feat(auth): simple auth scaffolding with types and endpoint installation (#22)
This PR scaffolds the types and endpoints required for four `/api/auth/...` endpoints, setting up the types and endpoint installation. The next PR will introduce a `Persistent` interface so that the endpoints actually do something, i.e. make queries to a SQL or SQLite database.

[See rendered documentation for more details](https://github.com/Originate/create-originate-app/blob/hl/auth/template/packages/auth/README.md)

The motivation behind this package was to refactor the common authentication scheme we implemented on Chief and Vibecheq into its own library, which had few dependencies on other libraries in the NPM ecosystem and was compatible with `@Originate/leash`.

## Side notes

* The package currently lives as a workspace of the `create-originate-app` template, but conceivably becomes its own repository in the future, at which point it could live as another GitHub package, like `@Originate/leash` does