backend: feat(crypto library) (#34)
This implements some basic crypto primitives for authentication:

- password digests (so we can store a salted, scrypted password in the database) 
- session token minting (using cryptographic signatures)

Care was taken to avoid the `crypto` library that comes with Node. That library has a dangerous interface for scrypt (takes a salt, even though scrypt algorithm is explicitly designed to generate a salt), and it lacks ed25519 signatures. Instead, we lean on two reliable libraries: `scrypt-kdf` and `tweetnacl`.